refactor: useItem fetch/list/buy 분리 - 코드 줄이 길어져서 가독성을 위해 분리 - purchase -> own으로 용어 변경 - myItems, mySaleItems 상태 관리 등록
