+++ src/app/mypage/transaction/components/TransactionProductList.tsx
@@ -1,38 +1,33 @@
 'use client';
 
-import React, { useEffect, useState } from 'react';
+import React, { useEffect } from 'react';
 import TransactionProductCard from './TransactionProductCard';
 import { useSearchParams } from 'next/navigation';
-import useItems from '@/hooks/useItems';
 import { convertToGifticonNFT, GifticonNFT } from '@/types';
 import LoadingOverlay from '@/components/ui/loadingSpinner';
 import { useItemStore } from '@/store/useItemStore';
 import { useAccount } from 'wagmi';
+import useFetchItems from '@/hooks/useFetchItems';
 
 export default function TransactionProductList() {
-	const [mySaleNFTs, setMySaleNFTs] = useState<GifticonNFT[]>([]);
-	const [myNFTs, setMyNFTs] = useState<GifticonNFT[]>([]);
 	const { address } = useAccount(); // ✅ 지갑 연결 상태 확인
-	const { items, setItems } = useItemStore();
+	const { items, myItems, mySaleItems, setItems, setMyItems, setMySaleItems } =
+		useItemStore();
 
 	const searchParams = useSearchParams();
 	const typeParam = searchParams.get('type') || 'own';
 	const type = typeParam === 'sale' ? 'sale' : 'own';
 
-	const { fetchItems, fetchMyNFTs, isLoading } = useItems();
+	const { fetchItems, fetchMyNFTs, isLoading } = useFetchItems();
 
 	useEffect(() => {
 		const fetch = async () => {
 			try {
-				const myNfts = await fetchMyNFTs();
-				console.log(
-					'nft.tokenIds:',
-					myNfts.map((nft) => nft.tokenId)
-				);
-				console.log(
-					'nft.tokenIds.toString():',
-					myNfts.map((nft) => nft.tokenId.toString())
-				);
+				if (myItems.length == 0) {
+					console.warn('🚨 내 아이템이 없습니다. 먼저 아이템을 불러오세요.');
+					const myItems = await fetchMyNFTs();
+					setMyItems(myItems);
+				}
 
 				if (items.length === 0) {
 					console.warn('🚨 아이템이 없습니다. 먼저 아이템을 불러오세요.');
@@ -50,8 +45,7 @@ export default function TransactionProductList() {
 					return acc;
 				}, []);
 
-				setMySaleNFTs(mySales);
-				setMyNFTs(myNfts);
+				setMySaleItems(mySales);
 			} catch (err) {
 				console.error('🚨 내 NFT 조회 중 오류:', err);
 			}
@@ -60,16 +54,14 @@ export default function TransactionProductList() {
 		fetch();
 	}, []);
 
-	const filteredNFTs = type === 'sale' ? mySaleNFTs : myNFTs;
+	const filteredNFTs = type === 'sale' ? mySaleItems : myItems;
 
 	if (isLoading) return <LoadingOverlay />;
 
 	if (filteredNFTs.length === 0) {
 		return (
 			<div className="p-8 text-center text-gray-500">
-				{type === 'own'
-					? '보유 중인 NFT가 없습니다.'
-					: '판매 내역이 없습니다.'}
+				{type === 'own' ? '보유 중인 NFT가 없습니다.' : '판매 내역이 없습니다.'}
 			</div>
 		);
 	}
